requirements.txt
Django==1.9.4
django-ckeditor==5.0.3
Pillow==3.2.0
six==1.10.0
